feat: refactor SimplePanel
