+++ src/components/PluginSettings/index.ts
@@ -1,0 +1,2 @@
+export * from "./TemplateEditor";
+export * from "./MultiSelectEditor";
