+++ src/module.ts
@@ -1,9 +1,10 @@
 import { PanelPlugin } from '@grafana/data';
-import { SimpleOptions } from './types';
-import { SimplePanel, SimpleEditor, TemplateEditor } from './components/SimplePanel';
+import { FlowPanel } from './components/SimplePanel';
+import { FlowOptions } from './types';
+import { TemplateEditor, MultiSelectEditor } from 'components/PluginSettings';
 
 
-export const plugin = new PanelPlugin<SimpleOptions>(SimplePanel).setPanelOptions((builder) => {
+export const plugin = new PanelPlugin<FlowOptions>(FlowPanel).setPanelOptions((builder) => {
     return builder
         .addCustomEditor({
             id: 'Template',
@@ -17,7 +18,7 @@ export const plugin = new PanelPlugin<SimpleOptions>(SimplePanel).setPanelOption
             name: 'Title',
             description: 'Title of flow item',
             defaultValue: '',
-            editor: SimpleEditor,
+            editor: MultiSelectEditor,
         })
         .addCustomEditor({
             id: 'aboveArrow',
@@ -25,7 +26,7 @@ export const plugin = new PanelPlugin<SimpleOptions>(SimplePanel).setPanelOption
             name: 'Above Arrow',
             description: 'Label directly above an arrow (hidden in simple format)',
             defaultValue: '',
-            editor: SimpleEditor,
+            editor: MultiSelectEditor,
         })
         .addCustomEditor({
             id: 'belowArrow',
@@ -33,7 +34,7 @@ export const plugin = new PanelPlugin<SimpleOptions>(SimplePanel).setPanelOption
             name: 'Below Arrow',
             description: 'Label directly below an arrow (hidden in simple format)',
             defaultValue: '',
-            editor: SimpleEditor,
+            editor: MultiSelectEditor,
         })
         .addCustomEditor({
             id: 'details',
@@ -41,7 +42,7 @@ export const plugin = new PanelPlugin<SimpleOptions>(SimplePanel).setPanelOption
             name: 'Details',
             description: 'Label 2 lines below an arrow',
             defaultValue: ['timestamp'],
-            editor: SimpleEditor,
+            editor: MultiSelectEditor,
         })
         .addCustomEditor({
             id: 'sourceLabel',
@@ -49,7 +50,7 @@ export const plugin = new PanelPlugin<SimpleOptions>(SimplePanel).setPanelOption
             name: 'Source Label',
             description: 'Label at the beginning of an arrow',
             defaultValue: '',
-            editor: SimpleEditor,
+            editor: MultiSelectEditor,
         })
         .addCustomEditor({
             id: 'destinationLabel',
@@ -57,14 +58,14 @@ export const plugin = new PanelPlugin<SimpleOptions>(SimplePanel).setPanelOption
             name: 'Destination Label',
             description: 'Label at the end of an arrow',
             defaultValue: '',
-            editor: SimpleEditor,
+            editor: MultiSelectEditor,
         })
         .addCustomEditor({
             id: 'source',
             path: 'source',
             name: 'Source Host',
             defaultValue: 'Label used to determine beginning of an arrow',
-            editor: SimpleEditor,
+            editor: MultiSelectEditor,
         })
         .addCustomEditor({
             id: 'destination',
@@ -72,7 +73,7 @@ export const plugin = new PanelPlugin<SimpleOptions>(SimplePanel).setPanelOption
             name: 'Destination Host',
             description: 'Label used to determine end of an arrow',
             defaultValue: '',
-            editor: SimpleEditor,
+            editor: MultiSelectEditor,
         })
         .addSelect({
             // id: 'sortoption',
@@ -93,7 +94,7 @@ export const plugin = new PanelPlugin<SimpleOptions>(SimplePanel).setPanelOption
             name: 'Items for color generator',
             description: 'Which labels should be used for background color generation',
             defaultValue: '',
-            editor: SimpleEditor,
+            editor: MultiSelectEditor,
         })
         .addBooleanSwitch({
             path: 'showbody',
