+++ src/types.ts
@@ -1,7 +1,15 @@
-type SeriesSize = 'sm' | 'md' | 'lg';
-
-export interface SimpleOptions {
-  text: string;
-  showSeriesCount: boolean;
-  seriesCountSize: SeriesSize;
+export interface FlowOptions {
+    title: OptionValue;
+    aboveArrow: OptionValue;
+    belowArrow: OptionValue;
+    details: OptionValue;
+    sourceLabel: OptionValue;
+    destinationLabel: OptionValue;
+    source: OptionValue;
+    destination: OptionValue;
+    sortoption: "none" | "time_old" | "time_new";
+    colorGenerator: OptionValue;
+    showbody: boolean;
 }
+
+type OptionValue = string | string[];
